Split out the code for the siblings table to a reusable component
