+++ src/app/artifacts-table/artifacts-table.component.css
@@ -1,12 +1,12 @@
 /* ----- Blocks ----- */
-.siblings-container {
+.artifacts-container {
     padding-left: 1.5rem;
     padding-right: 1.5rem;
     margin-bottom: 2.5rem;
 }
 
 /* A wrapper around the table for when it gets too wide */
-.siblings-table-responsive {
+.artifacts-table-responsive {
     overflow-x: auto;
     white-space: nowrap;
     /* Remove the default border provided by Bootstrap */
@@ -14,7 +14,7 @@
 }
 
 /* A wrapper around the table to create the section above the table for the "Show Columns" button to go */
-.siblings-table-wrapper {
+.artifacts-table-wrapper {
     background-color: #ffffff;
     display: inline-block;
     padding-top: 8rem; 
@@ -28,19 +28,19 @@
 }
 
 /* This div contains the "Show Columns" button. It is positioned in the DOM outside of
-   div.siblings-table-responsive, otherwise, it will cause a vertical scroll bar to appear
+   div.artifacts-table-responsive, otherwise, it will cause a vertical scroll bar to appear
    on short tables when the "Show Columns" button is expanded. */
-.siblings-table-header {
-    /* Make the width 100% minus the padding used by div.siblings-container */
+.artifacts-table-header {
+    /* Make the width 100% minus the padding used by div.artifacts-container */
     width: calc(100% - 3rem);
     /* Make the position absolute so the header is positioned on top of the padding-top of
-       div.siblings-table-wrapper */
+       div.artifacts-table-wrapper */
     position: absolute;
     margin-left: 2rem;
     margin-top: 2.4rem;
 }
 
-.siblings-table {
+.artifacts-table {
     margin: 0;
     font-size: 110%;
 }
@@ -55,7 +55,7 @@
 }
 
 @media (min-width: 992px) {
-    .siblings-container {
+    .artifacts-container {
         padding-left: 5rem;
         padding-right: 5rem;
         margin-bottom: 5rem;
@@ -66,35 +66,35 @@
         margin-bottom: 3rem;
     }
 
-    .siblings-table-header {
-        /* Make the width 100% minus the padding used by div.siblings-container */
+    .artifacts-table-header {
+        /* Make the width 100% minus the padding used by div.artifacts-container */
         width: calc(100% - 10rem);
     }
 }
 
 
 /* ----- Elements ----- */
-.siblings-table-header__btn-group {
+.artifacts-table-header__btn-group {
     /* A small hack to make the button vertically centered with the text on the line */
     position: relative;
     top: -3px;
 }
 
-.siblings-table-header__dropdown-button, .open > .siblings-table-header__dropdown-button,
-.siblings-table-header__dropdown-button:active {
+.artifacts-table-header__dropdown-button, .open > .artifacts-table-header__dropdown-button,
+.artifacts-table-header__dropdown-button:active {
     /* Override the colors default to Bootstrap */
     color: #363636;
     background-color: #ffffff;
     background-image: none;
 }
 
-.open > .siblings-table-header__dropdown-button {
+.open > .artifacts-table-header__dropdown-button {
     /* When the dropdown menu is open, don't round the bottom of the "Show Columns" button */
     border-bottom-left-radius: 0;
     border-bottom-right-radius: 0;
 }
 
-.siblings-table-header__dropdown-button {
+.artifacts-table-header__dropdown-button {
     font-size: 115%;
     /* Make the "Show Columns" button bigger */
     padding: 0.75rem 0.75rem 0.75rem 1rem;
@@ -103,12 +103,12 @@
     box-shadow: none;
 }
 
-.siblings-table-header__dropdown-button i {
+.artifacts-table-header__dropdown-button i {
     /* Give some space between the "Show Columns" text and the arrow icon on the button */
     padding-left: 2.5rem;
 }
 
-.siblings-table-header__dropdown-menu {
+.artifacts-table-header__dropdown-menu {
     padding-top: 2rem;
     padding-left: 2rem;
     margin-bottom: 2rem;
@@ -119,16 +119,16 @@
     overflow-x: hidden;
 }
 
-.siblings-table-header__dropdown-menu li {
+.artifacts-table-header__dropdown-menu li {
     font-size: 125%;
     padding-bottom: 1rem;
 }
 
-.siblings-table-header__dropdown-menu input {
+.artifacts-table-header__dropdown-menu input {
     margin-right: 0.75rem;
 }
 
-.siblings-table-header__columns-text {
+.artifacts-table-header__columns-text {
     /* Position this a little below where it normally is so it is centered with the button  */
     display: inline-block;
     margin-left: 1.5rem;
@@ -136,33 +136,33 @@
     font-size: 115%;
 }
 
-.siblings-table-header__columns-text-ending {
+.artifacts-table-header__columns-text-ending {
     /* Don't display the export to CSV button on mobile */
     display: none;
 }
 
-.siblings-table-header__export-btn {
+.artifacts-table-header__export-btn {
     /* Don't display the export to CSV button on mobile */
     display: none;
 }
 
-.siblings-table td, .siblings-table th {
+.artifacts-table td, .artifacts-table th {
     border: none;
     padding: 1rem;
 }
 
-.siblings-table th {
+.artifacts-table th {
     border-top: 1.5px solid #F2F2F2 !important;
     border-bottom: 1.5px solid #F2F2F2 !important;
 }
 
-.siblings-table th:hover {
+.artifacts-table th:hover {
     /* Make it seem like a link when hovering on the table header to let the user know they can sort */
     cursor: pointer;
     text-decoration: underline;
 }
 
-.siblings-table__sort-icon {
+.artifacts-table__sort-icon {
     padding-left: 0.5rem;
     /* Since the element is present in every column but the class that determines the icon
        is only active on sorted columns, we must set a constant width of the icon element
@@ -170,24 +170,24 @@
     width: 15px;
 }
 
-.siblings-table thead {
+.artifacts-table thead {
     /* Patternfly has a gradient background, so this removes it */
     background-image: none;
     font-size: 115%;
 }
 
-.siblings-table thead, .siblings-table tbody {
+.artifacts-table thead, .artifacts-table tbody {
     font-weight: 600;
     /* Set a background color on the table so that it doesn't blend with the background */
     background-color: #ffffff;
 }
 
 @media (min-width: 992px) {
-    .siblings-table-header__columns-text-ending {
+    .artifacts-table-header__columns-text-ending {
         display: initial;
     }
 
-    .siblings-table-header__export-btn {
+    .artifacts-table-header__export-btn {
         display: initial;
         margin-left: 3rem;
         font-size: 110%;
@@ -198,24 +198,24 @@
         color: #ffffff;
     }
 
-    .siblings-table-header__export-btn:hover {
+    .artifacts-table-header__export-btn:hover {
         background-color: #00659c;
     }
 
-    .siblings-table-header__export-btn:focus {
+    .artifacts-table-header__export-btn:focus {
         outline: none;
     }
 
-    .siblings-table {
-        /* Set the minimum width of the table to roughly that of div.siblings-table-header */
+    .artifacts-table {
+        /* Set the minimum width of the table to roughly that of div.artifacts-table-header */
         min-width: 500px;
     }
 
-    .siblings-table tr td:first-child, .siblings-table tr th:first-child {
+    .artifacts-table tr td:first-child, .artifacts-table tr th:first-child {
         padding-left: 2rem;
     }
 
-    .siblings-table tr td:last-child, .siblings-table tr th:last-child {
+    .artifacts-table tr td:last-child, .artifacts-table tr th:last-child {
         padding-right: 2rem;
     }
 }
