+++ src/app/artifacts-table/artifacts-table.component.html
@@ -1,0 +1,87 @@
+<div class="artifacts-container" *ngIf="formattedArtifacts">
+  <h2 class="title">{{ title }}</h2>
+  <div class="artifacts-table-header">
+      <!-- Set autoClose to false so that way the dropdown menu doesn't close when a user clicks on a check box -->
+      <div class="btn-group artifacts-table-header__btn-group" dropdown [autoClose]="false">
+        <button type="button" class="btn dropdown-toggle artifacts-table-header__dropdown-button" dropdownToggle>
+          Show Columns<i class="fa fa-angle-down"></i>
+        </button>
+        <!-- Create a dropdown menu with all the columns the user can choose to have shown -->
+        <ul *dropdownMenu class="dropdown-menu artifacts-table-header__dropdown-menu">
+          <li *ngFor="let column of columns | keys">
+            <!-- The default columns will be checked by default, any changes made by the user will trigger the
+                 setColumnState method to be called, which will store that state change in the columns component property. 
+                 This change will then be reflected by Angular when it reruns the various ngFors in this template. -->
+            <input type="checkbox" [checked]="columns[column]" (change)="setColumnState(column)" /> {{ column | propertyDisplay }}
+          </li>
+        </ul>
+      </div>
+      <div class="artifacts-table-header__columns-text">
+        {{ numActiveColumns }} of {{ (columns | keys).length }} columns
+        <span class="artifacts-table-header__columns-text-ending">selected</span>
+      </div>
+      <button class="artifacts-table-header__export-btn" (click)="exportToCsv()">Export CSV</button>
+  </div>
+  <div class="artifacts-table-responsive">
+    <div class="artifacts-table-wrapper">
+      <table class="table artifacts-table">
+        <thead>
+          <tr>
+            <!-- Display a header cell for every active column -->
+            <ng-container *ngFor="let column of columns | keys">
+              <th *ngIf="columns[column]" (click)="sortTable(column)">
+                <!-- Display an arrow if the table is being sorted by this column -->
+                {{ column | propertyDisplay }}<i class="fa artifacts-table__sort-icon"
+                    [ngClass]="{'fa-angle-down': sortedBy === column, 'fa-angle-up': sortedBy === '-' + column}"></i>
+              </th>
+            </ng-container>
+          </tr>
+        </thead>
+        <tbody>
+          <!-- Display a table row for every artifact -->
+          <tr *ngFor="let artifact of formattedArtifacts">
+            <!-- Display all the active columns for each artifact -->
+            <ng-container *ngFor="let column of columns | keys">
+              <td *ngIf="columns[column]">
+                <!-- If this column is the artifact's UID, make it a link to the corresponding external system -->
+                <ng-container *ngIf="uidColumns.includes(column); else nonUidProperty">
+                  <a href="{{ artifact | nodeExternalUrl }}" target="_blank">
+                    {{ artifact[column] }}
+                  </a>
+                </ng-container>
+                <ng-template #nonUidProperty>
+                  <ng-container *ngIf="(artifact[column]).length > 50; else shortProperty">
+                    <!-- If it's a long value, then truncate it in the table and make it so that when it is clicked,
+                         a modal is displayed with the whole text -->
+                    <a href="#" class="modal-link" (click)="openModal(truncateModal); $event.preventDefault()">
+                      {{ (artifact[column] | truncate: 50) }}
+                    </a>
+                    <!-- Define the modal template to be shown when the text is clicked -->
+                    <ng-template #truncateModal>
+                      <div class="modal-header">
+                        <h4 class="modal-title pull-left">{{ column | propertyDisplay }}</h4>
+                        <button type="button" class="close pull-right" (click)="modalRef.hide()">&times;</button>
+                      </div>
+                      <div class="modal-body">
+                        <ng-container *ngIf="preformattedColumns.includes(column); else normalValue">
+                          <pre>{{ artifact[column] }}</pre>
+                        </ng-container>
+                        <ng-template #normalValue>
+                          {{ artifact[column] }}
+                        </ng-template>
+                      </div>
+                    </ng-template>
+                  </ng-container>
+                  <!-- If the column is not a UID and the text is not too long, then just display it normally -->
+                  <ng-template #shortProperty>
+                      {{ artifact[column] }}
+                  </ng-template>
+                </ng-template>
+              </td>
+            </ng-container>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  </div>
+</div>
